requirements.txt
# Image-Processing lib
Pillow